+++ core/src/test/scala/org/locationtech/rasterframes/functions/AggregateFunctionsSpec.scala
@@ -1,0 +1,151 @@
+/*
+ * This software is licensed under the Apache 2 license, quoted below.
+ *
+ * Copyright 2019 Astraea, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ *     [http://www.apache.org/licenses/LICENSE-2.0]
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ */
+
+package org.locationtech.rasterframes.functions
+import geotrellis.raster._
+import geotrellis.raster.testkit.RasterMatchers
+import org.apache.spark.sql.Encoders
+import org.apache.spark.sql.functions._
+import org.locationtech.rasterframes._
+import org.locationtech.rasterframes.model.TileDimensions
+import org.locationtech.rasterframes.stats._
+import org.locationtech.rasterframes.tiles.ProjectedRasterTile.prtEncoder
+
+class AggregateFunctionsSpec extends TestEnvironment with RasterMatchers {
+  import TestData._
+  import spark.implicits._
+
+  implicit val pairEnc = Encoders.tuple(prtEncoder, prtEncoder)
+  implicit val tripEnc = Encoders.tuple(prtEncoder, prtEncoder, prtEncoder)
+
+  describe("aggregate statistics") {
+    it("should count data cells") {
+      val df = randNDTilesWithNull.filter(_ != null).toDF("tile")
+      df.select(rf_agg_data_cells($"tile")).first() should be(expectedRandData)
+      df.selectExpr("rf_agg_data_cells(tile)").as[Long].first() should be(expectedRandData)
+
+      checkDocs("rf_agg_data_cells")
+    }
+    it("should count no-data cells") {
+      val df = randNDTilesWithNull.toDF("tile")
+      df.select(rf_agg_no_data_cells($"tile")).first() should be(expectedRandNoData)
+      df.selectExpr("rf_agg_no_data_cells(tile)").as[Long].first() should be(expectedRandNoData)
+      checkDocs("rf_agg_no_data_cells")
+    }
+
+    it("should compute aggregate statistics") {
+      val df = randNDTilesWithNull.toDF("tile")
+
+      df.select(rf_agg_stats($"tile") as "stats")
+        .select("stats.data_cells", "stats.no_data_cells")
+        .as[(Long, Long)]
+        .first() should be((expectedRandData, expectedRandNoData))
+      df.selectExpr("rf_agg_stats(tile) as stats")
+        .select("stats.data_cells")
+        .as[Long]
+        .first() should be(expectedRandData)
+
+      checkDocs("rf_agg_stats")
+    }
+
+    it("should compute a aggregate histogram") {
+      val df = randNDTilesWithNull.toDF("tile")
+      val hist1 = df.select(rf_agg_approx_histogram($"tile")).first()
+      val hist2 = df
+        .selectExpr("rf_agg_approx_histogram(tile) as hist")
+        .select($"hist".as[CellHistogram])
+        .first()
+      hist1 should be(hist2)
+      checkDocs("rf_agg_approx_histogram")
+    }
+
+    it("should compute local statistics") {
+      val df = randNDTilesWithNull.toDF("tile")
+      val stats1 = df
+        .select(rf_agg_local_stats($"tile"))
+        .first()
+      val stats2 = df
+        .selectExpr("rf_agg_local_stats(tile) as stats")
+        .select($"stats".as[LocalCellStatistics])
+        .first()
+
+      stats1 should be(stats2)
+      checkDocs("rf_agg_local_stats")
+    }
+
+    it("should compute local min") {
+      val df = Seq(two, three, one, six).toDF("tile")
+      df.select(rf_agg_local_min($"tile")).first() should be(one.toArrayTile())
+      df.selectExpr("rf_agg_local_min(tile)").as[Tile].first() should be(one.toArrayTile())
+      checkDocs("rf_agg_local_min")
+    }
+
+    it("should compute local max") {
+      val df = Seq(two, three, one, six).toDF("tile")
+      df.select(rf_agg_local_max($"tile")).first() should be(six.toArrayTile())
+      df.selectExpr("rf_agg_local_max(tile)").as[Tile].first() should be(six.toArrayTile())
+      checkDocs("rf_agg_local_max")
+    }
+
+    it("should compute local mean") {
+      checkDocs("rf_agg_local_mean")
+      val df = Seq(two, three, one, six)
+        .toDF("tile")
+        .withColumn("id", monotonically_increasing_id())
+
+      df.select(rf_agg_local_mean($"tile")).first() should be(three.toArrayTile())
+
+      df.selectExpr("rf_agg_local_mean(tile)").as[Tile].first() should be(three.toArrayTile())
+
+      noException should be thrownBy {
+        df.groupBy($"id")
+          .agg(rf_agg_local_mean($"tile"))
+          .collect()
+      }
+    }
+
+    it("should compute local data cell counts") {
+      val df = Seq(two, randNDPRT, nd).toDF("tile")
+      val t1 = df.select(rf_agg_local_data_cells($"tile")).first()
+      val t2 = df.selectExpr("rf_agg_local_data_cells(tile) as cnt").select($"cnt".as[Tile]).first()
+      t1 should be(t2)
+      checkDocs("rf_agg_local_data_cells")
+    }
+
+    it("should compute local no-data cell counts") {
+      val df = Seq(two, randNDPRT, nd).toDF("tile")
+      val t1 = df.select(rf_agg_local_no_data_cells($"tile")).first()
+      val t2 = df.selectExpr("rf_agg_local_no_data_cells(tile) as cnt").select($"cnt".as[Tile]).first()
+      t1 should be(t2)
+      val t3 = df.select(rf_local_add(rf_agg_local_data_cells($"tile"), rf_agg_local_no_data_cells($"tile"))).as[Tile].first()
+      t3 should be(three.toArrayTile())
+      checkDocs("rf_agg_local_no_data_cells")
+    }
+  }
+
+  describe("aggregate rasters") {
+    it("should create a global aggregate raster from projected raster column") {
+      val df = rgbCogSample.toDF(TileDimensions(32, 32))
+    //  df.agg(rf_agg_overview_raster(500, 400, df.tileColumns: _*))
+      df.show(false)
+    }
+  }
+}